Update serialization utils
